pyproject.toml
[project]
name = "fund-screener"
version = "1.2.0"
description = "基金筛选系统 - 自用基金筛选工具，重点筛选稳健增长、长期复利的基金"
readme = "README.md"
requires-python = ">=3.9"
license = { text = "MIT" }
authors = [
    { name = "Claude", email = "claude@example.com" }
]
keywords = ["fund", "investment", "screener", "quant"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "pandas>=2.0.0",
    "numpy>=1.24.0",
    "sqlalchemy>=2.0.0",
    "pymysql>=1.1.0",
    # "psycopg2-binary>=2.9.0",  # PostgreSQL (保留但未启用)
    "alembic>=1.12.0",
    "akshare",
    "requests>=2.31.0",
    "apscheduler>=3.10.0",
    "python-dotenv>=1.0.0",
    "loguru>=0.7.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]

[project.scripts]
fund-screener = "fund_screener.__main__:main"

[project.urls]
Homepage = "https://github.com/example/fund-screener"
Repository = "https://github.com/example/fund-screener"
Documentation = "https://github.com/example/fund-screener#readme"

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]

[dependency-groups]
dev = [
    "pytest>=7.4.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]

[tool.black]
line-length = 88
target-version = ['py39', 'py310', 'py311', 'py312']
include = '\.pyi?$'

[tool.ruff]
line-length = 88
target-version = "py39"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]
ignore = ["E501"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
